[diffusion] CI: fix generate mode and add cli test
