+++ python/sglang/multimodal_gen/test/cli/test_generate_common.py
@@ -3,105 +3,187 @@
 """
     Common generate cli test, one test for image and video each
 """
-
+import dataclasses
 import os
+import shlex
+import subprocess
+import sys
 import unittest
-from pathlib import Path
+from typing import Optional
 
 from PIL import Image
 
+from sglang.multimodal_gen.configs.sample.sampling_params import DataType
+from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 from sglang.multimodal_gen.test.test_utils import (
-    TestCLIBase,
     check_image_size,
-    is_mp4,
-    run_command,
 )
 
+logger = init_logger(__name__)
 
-class TestGenerate(TestCLIBase):
-    model_path = "black-forest-labs/FLUX.1-dev"
-    launch_file_name = "launch_flux.json"
-    output_name = "FLUX.1-dev, single gpu"
-    ext = "jpg"
 
-    def test_generate_with_config(self):
-        test_dir = Path(__file__).parent
-        config_path = (
-            (test_dir / ".." / "test_files" / self.launch_file_name)
-            .resolve()
-            .as_posix()
+
+
+@dataclasses.dataclass
+class TestResult:
+    name: str
+    key: str
+    duration: Optional[float]
+    succeed: bool
+
+    @property
+    def duration_str(self):
+        return f"{self.duration:.4f}" if self.duration else "NA"
+
+
+
+def run_command(command) -> Optional[float]:
+    """Runs a command and returns the execution time and status."""
+    print(f"Running command: {shlex.join(command)}")
+
+    duration = None
+    with subprocess.Popen(
+        command,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        text=True,
+        encoding="utf-8",
+    ) as process:
+        for line in process.stdout:
+            sys.stdout.write(line)
+            if "Pixel data generated" in line:
+                words = line.split(" ")
+                duration = float(words[-2])
+
+    if process.returncode == 0:
+        return duration
+    else:
+        print(f"Command failed with exit code {process.returncode}")
+        return None
+
+
+class TestCLIBase(unittest.TestCase):
+    model_path: str = None
+    extra_args = []
+    data_type: DataType = None
+    # tested on h100
+    thresholds = {}
+
+    width: int = 720
+    height: int = 720
+    output_path: str = "test_outputs"
+
+    base_command = [
+        "sglang",
+        "generate",
+        "--text-encoder-cpu-offload",
+        "--pin-cpu-memory",
+        "--prompt",
+        "A curious raccoon",
+        "--save-output",
+        "--log-level=debug",
+        f"--width={width}",
+        f"--height={height}",
+        f"--output-path={output_path}",
+    ]
+
+    results = []
+
+    @classmethod
+    def setUpClass(cls):
+        cls.results = []
+
+    def _run_command(self, name: str, model_path: str, test_key: str = "", args=[]):
+        command = (
+            self.base_command
+            + [f"--model-path={model_path}"]
+            + shlex.split(args or "")
+            + ["--output-file-name", f"{name}"]
+            + self.extra_args
         )
-        command = [
-            "sgl_diffusion",
-            "generate",
-            f"--config={config_path}",
-        ]
         duration = run_command(command)
+        status = "Success" if duration else "Failed"
+        succeed = duration is not None
 
-        self.assertIsNotNone(duration, f"Run command failed: {command}")
+        duration = float(duration) if succeed else None
+        self.results.append(TestResult(name, test_key, duration, succeed))
 
-        # verify
-        self.verify_image(self.output_name)
-
-    def test_generate_multiple_outputs(self):
-        command = [
-            "sglang",
-            "generate",
-            "--prompt",
-            "A curious raccoon",
-            "--output-path=outputs",
-            f"--model-path={self.model_path}",
-            "--save-output",
-            f"--output-file-name={self.output_name}",
-            "--num-outputs-per-prompt=2",
-            "--width=720",
-            "--height=720",
-        ]
-        duration = run_command(command)
-        self.assertIsNotNone(duration, f"Run command failed: {command}")
-
-        self.verify_image(f"{self.output_name}_0.{self.ext}")
-        self.verify_image(f"{self.output_name}_1.{self.ext}")
-
-    def verify_image(self, output_name):
-        path = os.path.join("outputs", output_name)
-        with Image.open(path) as image:
-            check_image_size(self, image, 720, 720)
-
-    def verify_video(self, output_name):
-        path = os.path.join("outputs", output_name)
-        with open(path, "rb") as f:
-            header = f.read(12)
-            assert is_mp4(header)
+        return name, duration, status
 
 
-class TestWanGenerate(TestGenerate):
-    model_path = "Wan-AI/Wan2.1-T2V-1.3B-Diffusers"
-    launch_file_name = "launch_wan.json"
-    output_name = "Wan2.1-T2V-1.3B-Diffusers, single gpu"
-    ext = "mp4"
+class TestGenerateBase(TestCLIBase):
+    model_path: str = None
+    extra_args = []
+    data_type: DataType = None
+    # tested on h100
+    thresholds = {}
 
-    def test_generate_multiple_outputs(self):
-        command = [
-            "sglang",
-            "generate",
-            "--prompt",
-            "A curious raccoon",
-            "--output-path=outputs",
-            f"--model-path={self.model_path}",
-            "--save-output",
-            f"--output-file-name={self.output_name}",
-            "--num-outputs-per-prompt=2",
-            "--width=720",
-            "--height=720",
-        ]
-        duration = run_command(command)
-        self.assertIsNotNone(duration, f"Run command failed: {command}")
+    width: int = 720
+    height: int = 720
+    output_path: str = "test_outputs"
+    image_path: str | None = None
+    prompt: str | None = "A curious raccoon"
 
-        self.verify_video(f"{self.output_name}_0.{self.ext}")
-        # FIXME: second video is a meaningless output
-        self.verify_video(f"{self.output_name}_1.{self.ext}")
+    base_command = [
+        "sglang",
+        "generate",
+        f"--prompt",
+        f"{prompt}",
+        "--save-output",
+        "--log-level=debug",
+        f"--width={width}",
+        f"--height={height}",
+        f"--output-path={output_path}",
+    ]
 
+    results: list[TestResult] = []
 
-if __name__ == "__main__":
-    unittest.main()
+    @classmethod
+    def setUpClass(cls):
+        cls.results = []
+
+    @classmethod
+    def tearDownClass(cls):
+        pass
+
+    def _run_test(self, name: str, args, model_path: str, test_key: str):
+        time_threshold = self.thresholds[test_key]
+        name, duration, status = self._run_command(
+            name, args=args, model_path=model_path, test_key=test_key
+        )
+        self.verify(status, name, duration, time_threshold)
+
+    def verify(self, status, name, duration, time_threshold):
+        print("-" * 80)
+        print("\n" * 3)
+
+        # test task status
+        self.assertEqual(status, "Success", f"{name} command failed")
+        self.assertIsNotNone(duration, f"Could not parse duration for {name}")
+        self.assertLessEqual(
+            duration,
+            time_threshold,
+            f"{name} failed with {duration:.4f}s > {time_threshold}s",
+        )
+
+        # test output file
+        path = os.path.join(
+            self.output_path, f"{name}.{self.data_type.get_default_extension()}"
+        )
+        self.assertTrue(os.path.exists(path), f"Output file not exist for {path}")
+        if self.data_type == DataType.IMAGE:
+            with Image.open(path) as image:
+                check_image_size(self, image, self.width, self.height)
+        logger.info(f"{name} passed in {duration:.4f}s (threshold: {time_threshold}s)")
+
+    def model_name(self):
+        return self.model_path.split("/")[-1]
+
+    def test_single_gpu(self):
+        """single gpu"""
+        self._run_test(
+            name=f"{self.model_name()}_single_gpu",
+            args=None,
+            model_path=self.model_path,
+            test_key="test_single_gpu",
+        )
